Fixed loading of styles, hover works in searchresults

+++ javascript/forumloadlink.js
@@ -51,9 +51,19 @@ document.addEventListener("DOMContentLoaded", () => {
     // Add eventlistener to each sidebar button to call moveToForum() when clicked
     for(let i = 0; i < sideBarButtons.length; i++)
         sideBarButtons[i].addEventListener("click", () => {
-    
-            moveToForum(sideBarButtons[i].textContent.trim());
-            sessionStorage.clear();
+           
+            if(sideBarButtons[i].textContent.trim() === "Rules") {
+                moveToForum("Rules and Regulations")
+            }
+            if(sideBarButtons[i].textContent.trim() === "Announcements") {
+            
+                moveToForum("Forum Announcements")
+            }
+            else{
+                moveToForum(sideBarButtons[i].textContent.trim());
+                sessionStorage.clear();
+            }
+
     
         });
 
@@ -101,13 +111,11 @@ document.addEventListener("DOMContentLoaded", () => {
         // Switch statement obtains the needed information depending on the type parameter
         switch (type){
 
-            case "Announcements":
             case "Forum Announcements":
                 forumImage = "../resources/images/announcements.jpg";
                 forumDescription = "Learn important information about Blevvit.";
                 break;
 
-            case "Rules": 
             case "Rules and Regulations": 
                 forumImage = "../resources/images/rules_and_regulations.jpg";
                 forumDescription = "Learn how to be a respectful member in Blevvit.";
